+++ test/registered/debug_utils/comparator/unshard/test_execute.py
@@ -4,8 +4,8 @@ import pytest
 import torch
 
 from sglang.srt.debug_utils.comparator.dims import parse_dims
-from sglang.srt.debug_utils.comparator.unshard.execute import execute_unshard_plan
-from sglang.srt.debug_utils.comparator.unshard.plan import compute_unshard_plan
+from sglang.srt.debug_utils.comparator.unshard.executor import execute_unshard_plan
+from sglang.srt.debug_utils.comparator.unshard.planner import compute_unshard_plan
 from sglang.srt.debug_utils.comparator.unshard.types import AxisInfo, UnshardPlan
 from sglang.test.ci.ci_register import register_cpu_ci
